Refactor components out into their own files
