+++ packages/web/src/components/logo.tsx
@@ -1,0 +1,7 @@
+import LogoImage from "@/assets/dddforumlogo.png";
+
+export const Logo = () => (
+  <div>
+    <img className="h-14" src={LogoImage} alt="DDD Forum Logo"/>
+  </div>
+);
